+++ src/app/modules/stepper/stepper.component.spec.ts
@@ -56,6 +56,7 @@ class DefineCohortPageComponentStub {
 
 describe('StepperComponent', () => {
   let component: StepperComponent;
+  let fhirBackend: FhirBackendService;
   let fixture: ComponentFixture<StepperComponent>;
 
   beforeEach(async () => {
@@ -81,14 +82,15 @@ describe('StepperComponent', () => {
         {
           provide: FhirBackendService,
           useValue: {
-            initialized: new BehaviorSubject(ConnectionStatus.Ready)
+            initialized: new BehaviorSubject(ConnectionStatus.Ready),
+            features: { batch: true },
+            disconnect: () => {}
           }
         },
         {
           provide: ColumnDescriptionsService,
           useValue: {
-            getVisibleColumns: () => of([]),
-            destroy: () => {}
+            getVisibleColumns: () => of([])
           }
         }
       ]
@@ -96,24 +98,28 @@ describe('StepperComponent', () => {
   });
 
   beforeEach(async () => {
+    fhirBackend = TestBed.inject(FhirBackendService);
+    spyOn(fhirBackend, 'disconnect').and.callThrough();
     fixture = TestBed.createComponent(StepperComponent);
     component = fixture.componentInstance;
     await fixture.detectChanges();
     spyOn(component.subscription, 'unsubscribe').and.callThrough();
-    spyOn(component.columnDescriptions, 'destroy').and.callThrough();
   });
 
   it('should create', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should initialize', () => {
-    expect(component.serverInitialized).toBe(true);
+  it('should initialize correctly', () => {
+    component.stepper.steps.forEach((step) => {
+      // Only the first step is completed
+      expect(step.completed).toBe(step === component.stepper.steps.first);
+    });
   });
 
   it('should unsubscribe on destroy', () => {
     component.ngOnDestroy();
     expect(component.subscription.unsubscribe).toHaveBeenCalledOnceWith();
-    expect(component.columnDescriptions.destroy).toHaveBeenCalledOnceWith();
+    expect(fhirBackend.disconnect).toHaveBeenCalledOnceWith();
   });
 });